Page allUsers.html.twig and allBaskets.html.twig done

+++ src/Controller/BasketController.php
@@ -14,9 +14,14 @@ use Symfony\Component\Routing\Annotation\Route;
 #[Route('{_locale}/basket')]
 class BasketController extends AbstractController
 {
+
     #[Route('/', name: 'app_basket_index', methods: ['GET'])]
     public function index(BasketRepository $basketRepository, BasketContentRepository $basketContentRepository): Response
     {
+        if(!$this->getUser()){
+            return $this->redirectToRoute('app_login');
+        }
+
         return $this->render('basket/index.html.twig', [
             'baskets' => $basketRepository->findByUtilisateur($this->getUser()),
             'basket_contents' => $basketContentRepository->findByBasket('basket'),
@@ -26,6 +31,10 @@ class BasketController extends AbstractController
     #[Route('/allBaskets', name: 'app_all_basket_index', methods: ['GET'])]
     public function indexAll(BasketRepository $basketRepository): Response
     {
+        if(!$this->getUser()){
+            return $this->redirectToRoute('app_login');
+        }
+        
         return $this->render('basket/allBaskets.html.twig', [
             'baskets' => $basketRepository->findAll(),
         ]);
@@ -33,9 +42,13 @@ class BasketController extends AbstractController
 
 
 
-    #[Route('/new', name: 'app_basket_new', methods: ['GET', 'POST'])]
+    #[Route('basket/action/new', name: 'app_basket_new', methods: ['GET', 'POST'])]
     public function new(Request $request, BasketRepository $basketRepository): Response
     {
+        if(!$this->getUser()){
+            return $this->redirectToRoute('app_login');
+        }
+
         $basket = new Basket();
         $form = $this->createForm(BasketType::class, $basket);
         $form->handleRequest($request);
@@ -51,17 +64,24 @@ class BasketController extends AbstractController
         ]);
     }
 
-    #[Route('/{id}', name: 'app_basket_show', methods: ['GET'])]
+    #[Route('/basket/{id}', name: 'app_basket_show', methods: ['GET'])]
     public function show(Basket $basket): Response
     {
+        if(!$this->getUser()){
+            return $this->redirectToRoute('app_login');
+        }
+
         return $this->render('basket/show.html.twig', [
             'basket' => $basket,
         ]);
     }
-
-    #[Route('/{id}/edit', name: 'app_basket_edit', methods: ['GET', 'POST'])]
+    #[Route('basket/action/{id}/edit', name: 'app_basket_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Basket $basket, BasketRepository $basketRepository): Response
     {
+        if(!$this->getUser()){
+            return $this->redirectToRoute('app_login');
+        }
+
         $form = $this->createForm(BasketType::class, $basket);
         $form->handleRequest($request);
 
@@ -71,15 +91,19 @@ class BasketController extends AbstractController
             return $this->redirectToRoute('app_basket_index', [], Response::HTTP_SEE_OTHER);
         }
 
-        return $this->renderForm('basket/edit.html.twig', [
+        return $this->render('basket/edit.html.twig', [
             'basket' => $basket,
             'form' => $form,
         ]);
     }
 
-    #[Route('/{id}', name: 'app_basket_delete', methods: ['POST'])]
+    #[Route('/basket/action/{id}', name: 'app_basket_delete', methods: ['POST'])]
     public function delete(Request $request, Basket $basket, BasketRepository $basketRepository): Response
     {
+        if(!$this->getUser()){
+            return $this->redirectToRoute('app_login');
+        }
+
         if ($this->isCsrfTokenValid('delete'.$basket->getId(), $request->request->get('_token'))) {
             $basketRepository->remove($basket, true);
         }
